Make search highlighting still work correctly when recursion is flattened
