+++ src/views/application-container.tsx
@@ -1,10 +1,10 @@
 import {h} from 'preact'
-import {Application, ActiveProfileState} from './application'
-import {getProfileToView, getCanvasContext} from '../store/getters'
+import {Application} from './application'
+import {getCanvasContext} from '../store/getters'
 import {actions} from '../store/actions'
 import {useActionCreator} from '../lib/preact-redux'
 import {memo} from 'preact/compat'
-import {useAppSelector} from '../store'
+import {useAppSelector, useActiveProfileState} from '../store'
 
 const {
   setLoading,
@@ -24,26 +24,9 @@ export const ApplicationContainer = memo(() => {
     [],
   )
 
-  const activeProfileState: ActiveProfileState | null = useAppSelector(state => {
-    const {profileGroup} = state
-    if (!profileGroup) return null
-    if (profileGroup.indexToView >= profileGroup.profiles.length) return null
-
-    const index = profileGroup.indexToView
-    const profileState = profileGroup.profiles[index]
-    return {
-      ...profileGroup.profiles[profileGroup.indexToView],
-      profile: getProfileToView({
-        profile: profileState.profile,
-        flattenRecursion: state.flattenRecursion,
-      }),
-      index: profileGroup.indexToView,
-    }
-  }, [])
-
   return (
     <Application
-      activeProfileState={activeProfileState}
+      activeProfileState={useActiveProfileState()}
       canvasContext={canvasContext}
       setGLCanvas={useActionCreator(setGLCanvas, [])}
       setLoading={useActionCreator(setLoading, [])}
